requirements.txt
requests==2.32.0
python-dotenv~=1.0.1
allure-python-commons~=2.13.5
pages~=0.3
webdriver-manager~=4.0.2
selenium==4.20.0
SQLAlchemy==1.4.50
allure-pytest==2.13.5
pytest==8.2.0
autopep8==2.2.0
charset-normalizer==3.3.2
flake8==7.0.0
iniconfig==2.0.0
packaging==24.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.2.0
PySocks==1.7.1
